Merge dev into master

+++ aws/ec2/ec2_stop_instance.sh
@@ -5,7 +5,7 @@
 # Usage:
 #     Run script with --help option to get usage.
 
-version="1.0.9"
+version="1.1.1"
 
 set -o pipefail
 script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
@@ -16,12 +16,13 @@ usage() {
     echo "Usage:"
     echo "    export AWS_PROFILE=profile"
     echo
-    echo "    $script_name [-f] [-w] {name | instance_id}..."
+    echo "    $script_name [-f] [-n] [-w] {name | instance_id}..."
     echo
     echo "Description:"
-    echo "    -f, --force   Do not prompt for confirmation."
-    echo "    -w, --wait    Wait for 'stopped' state before finishing."
-    echo "    -h, --help    Display this help."
+    echo "    -f, --force        Force stop instance."
+    echo "    -n, --no-prompt    No confirmation prompt."
+    echo "    -w, --wait         Wait for 'stopped' state before finishing."
+    echo "    -h, --help         Display this help."
     exit 1
 }
 
@@ -31,6 +32,10 @@ while test -n "$1"; do
         usage
         ;;
     -f|--force)
+        force="yes"
+        shift
+        ;;
+    -n|--no-prompt)
         do_not_prompt="yes"
         shift
         ;;
@@ -83,7 +88,7 @@ for name in $names; do
 
     aws --profile "$profile" ec2 describe-volumes --filters "Name=attachment.instance-id, Values=$instance_id" --query 'Volumes[*].{VolumeId:VolumeId,InstanceId:Attachments[0].InstanceId,State:Attachments[0].State,DeleteOnTermination:Attachments[0].DeleteOnTermination,Device:Attachments[0].Device,Size:Size}' --output table | tee -a "$log"
 
-    if [[ $do_not_prompt != "yes" ]]; then
+    if [[ ! $do_not_prompt ]]; then
         echo -n 'Are you sure that you want this instance stopped? y/n: '
         read -r yn
         if [[ $yn != y ]]; then
@@ -94,7 +99,11 @@ for name in $names; do
     fi
 
     echo "Stop instance..." | tee -a "$log"
-    aws --profile "$profile" ec2 stop-instances --instance-ids "$instance_id" --output table | tee -a "$log"
+    if [[ $force ]]; then
+        aws --profile "$profile" ec2 stop-instances --instance-ids "$instance_id" --force --output table | tee -a "$log"
+    else
+        aws --profile "$profile" ec2 stop-instances --instance-ids "$instance_id" --output table | tee -a "$log"
+    fi
     state=""
     while [[ $state != "stopped" ]] && [[ $wait_for_stopped == yes ]]; do
         state=$(aws --profile "$profile" ec2 describe-instances --instance-ids "$instance_id" --query 'Reservations[].Instances[].[State.Name]' --output text)
